Torch export Integration (#996)
* Enforcing parameter and buffer tieing between dmx module and original module

Adding prototype for torch export

working prototype

adding example with legacy cache

Updating demo with comments

testing and adding more HF models in export prototype

Adding prototype for dynamo export

adding option to use dynamo.export for DmxModel.from_torch

Reusing of submodules from the already existing _gms

adding method substitution

Adding prototype for torch export unflatten

Adding submod forward verification in prototype

Adding recursive substitution transformer for unflattened exported module

Adding function substitution for non compound functions

Migrating to export: opt with sdpa fully working

Adding test case

fixing minor issue with parameter dtype

aligning kwargs order with model signature

Adding dtype alignment for smoothquant and device alignment for gptq

Turning BaseModelOutput inputs to tuple: Whisper working

Process inputs for modules not packed by unflatten: eg, transformers.Conv1D: gpt2 working

Adding missing ops: CLIP working

Throw ValueError for accelerate dispatched models since they are not supported by export

Cleanup and adding test

Adding additonal mapping support for export path

Removing prototype examples

Preparing preserve_module_call_signature for export

qwen3 working with custom RMSNorm and RotaryEmbedding

qwen forward pass working

* Fixing rebase

* Adding unit tests

* Adding examples for torch export usage

* Fixing failing test cases

* Making apply_rotary_pos_emb a custom op

* adding some checks of versioning for backward compatibility

* cleanup

* Adding custom ops file

* renaming custom op library

+++ src/dmx/compressor/numerical/smoothquant.py
@@ -482,7 +482,7 @@ class ActivationWeightSmoothQuant(SmoothQuant):
         Returns:
             scaled weight tensor
         """
-        return self.scale_b(wgt).to(wgt.device)
+        return self.scale_b(wgt).to(wgt.device).to(wgt.dtype)
 
     def scale_input(self, inp):
         """
